Modif doc and add comment

+++ hugging_face/app.py
@@ -102,8 +102,9 @@ def query_image(image):
             predict.append(
                 {
                     # The list of coordinates of the points of the polygon.
-                    # Use np.asarray to change list to ndarray to turn int64 to int32 with np.ndarray.tolist
-                    "polygon": np.asarray(polygon["polygon"]).tolist(),
+                    # Use np.asarray to change list to ndarray to transform int64 to int32 with
+                    # np.ndarray.tolist to be able to serialize it to JSON.
+                    "polygon": np.asarray(polygon["polygon"], dtype=np.int32).tolist(),
                     # Confidence that the model predicts the polygon in the right place
                     "confidence": polygon["confidence"],
                     # The channel on which the polygon is predicted
